+++ funciones/src/main/java/com/veterinaria/dto/MascotaRequestDTO.java
@@ -1,15 +1,15 @@
 package com.veterinaria.dto;
 
 public record MascotaRequestDTO(
-                Long id,
-                Long clienteId,
-                String nombre,
-                Integer especieId,
-                Integer razaId,
-                String fechaNacimiento, // ISO yyyy-MM-dd
-                Integer sexoId,
-                Integer estadoId,
-                String fotoUrl,
-                String otraEspecie,
-                String otraRaza) {
+        Long id,
+        Long clienteId,
+        String nombre,
+        Long especieId,
+        Long razaId,
+        String fechaNacimiento, // ISO yyyy-MM-dd
+        Long sexoId,
+        Long estadoId,
+        String fotoUrl,
+        String otraEspecie,
+        String otraRaza) {
 }
